Lowering stress test to not blow up travis
