+++ core/src/main/java/io/paradoxical/cassieq/factories/DataContextFactory.java
@@ -10,3 +10,4 @@ public interface DataContextFactory {
 
     Optional<QueueDefinition> getDefinition(QueueName name);
 }
+
