+++ core/src/main/java/io/paradoxical/cassieq/workers/reader/InvisLocator.java
@@ -1,0 +1,4 @@
+package io.paradoxical.cassieq.workers.reader;
+
+public interface InvisLocator {
+}
